move the separate index
