+++ src/libs/trpc/lambda/index.ts
@@ -1,0 +1,1 @@
+export * from './serverDatabase';
