+++ src/server/routers/lambda/agent.ts
@@ -8,7 +8,8 @@ import { KnowledgeBaseModel } from '@/database/models/knowledgeBase';
 import { SessionModel } from '@/database/models/session';
 import { UserModel } from '@/database/models/user';
 import { pino } from '@/libs/logger';
-import { authedProcedure, router, serverDatabase } from '@/libs/trpc';
+import { authedProcedure, router } from '@/libs/trpc';
+import { serverDatabase } from '@/libs/trpc/lambda';
 import { AgentService } from '@/server/services/agent';
 import { KnowledgeItem, KnowledgeType } from '@/types/knowledgeBase';
