+++ src/server/routers/lambda/message.ts
@@ -3,7 +3,8 @@ import { z } from 'zod';
 import { MessageModel } from '@/database/models/message';
 import { updateMessagePluginSchema } from '@/database/schemas';
 import { getServerDB } from '@/database/server';
-import { authedProcedure, publicProcedure, router, serverDatabase } from '@/libs/trpc';
+import { authedProcedure, publicProcedure, router } from '@/libs/trpc';
+import { serverDatabase } from '@/libs/trpc/lambda';
 import { getFullFileUrl } from '@/server/utils/files';
 import { ChatMessage } from '@/types/message';
 import { BatchTaskResult } from '@/types/service';
