+++ src/server/routers/lambda/plugin.ts
@@ -2,7 +2,8 @@ import { z } from 'zod';
 
 import { PluginModel } from '@/database/models/plugin';
 import { getServerDB } from '@/database/server';
-import { authedProcedure, publicProcedure, router, serverDatabase } from '@/libs/trpc';
+import { authedProcedure, publicProcedure, router } from '@/libs/trpc';
+import { serverDatabase } from '@/libs/trpc/lambda';
 import { LobeTool } from '@/types/tool';
 
 const pluginProcedure = authedProcedure.use(serverDatabase).use(async (opts) => {
